Cloud/requirements.txt
# Function dependencies, for example:
# package>=version
Keras==2.4.3
matplotlib==3.2.2
numpy==1.19.5
pandas==1.1.5
tensorflow==2.5.0
google-cloud-storage==1.16.1
librosa==0.8.0
google-cloud-firestore == 2.1.1
google-api-python-client==2.7.0